pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-bench"
version = "0.1.0"
description = "Benchmark any LLM provider against your actual prompts — latency, cost, quality"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Midnight Run", email = "hi@midnightrun.ai" }
]
keywords = [
    "llm", "benchmark", "openai", "anthropic", "gemini", "mistral", "groq",
    "cli", "ai", "evaluation", "latency", "cost"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Testing",
    "Topic :: Utilities",
]

dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.30"]
gemini = ["google-generativeai>=0.8"]
mistral = ["mistralai>=1.0"]
groq = ["groq>=0.9"]
all = [
    "openai>=1.0",
    "anthropic>=0.30",
    "google-generativeai>=0.8",
    "mistralai>=1.0",
    "groq>=0.9",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.0",
    "hatch",
    "twine",
]

[project.urls]
Homepage = "https://midnightrun.ai/llm-bench"
Repository = "https://github.com/midnightrunai/llm-bench"
Issues = "https://github.com/midnightrunai/llm-bench/issues"
Changelog = "https://github.com/midnightrunai/llm-bench/blob/main/CHANGELOG.md"

[project.scripts]
llm-bench = "llm_bench.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["llm_bench"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
